requirements.txt
flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
twilio==7.0.0
supabase==2.15.0
mysql-connector-python==8.0.26 